update to smali 2.0.3

+++ brut.apktool.smali/dexlib2/src/main/java/org/jf/dexlib2/iface/instruction/formats/UnknownInstruction.java
@@ -32,5 +32,5 @@
 package org.jf.dexlib2.iface.instruction.formats;
 
 public interface UnknownInstruction extends Instruction10x {
-    short getOriginalOpcode();
+    int getOriginalOpcode();
 }
